Added explicit support for x86_64
Previously, we just relied on implicit support given that x86 is compatible with both 32 bit and 64 bit systems but going forward, the Play Store will reject these builds, EVEN FOR INTERAL TESTS. This addresses that but it comes with other considerations. arm64 and x86_64 are only supported on Android for API 21+. We probably should structure our resource folders differently to reflect this nuance. Our supported versions are: API 16+ for 32-bit hardware and API 21+ for 64-bit hardware, unless its 64-bit arm which is backward compatible with our 32-bit arm code. Clear as mud, right? I'll add this verbiage somewhere.