Merge initializers.
The new API for initializing with a viewing key will support all existing uses of the other initializer. So they can be safely merged and the changes to support the new one are fairly straight forward.

Also, iterated on the design of the Initializer to simplify construction and make the API easier to use.