New: Shorten the UTXO retry time.
This is temporary PoC logic to not constantly look for UTXOs on a lightwalletd instance that does not seem to support t-addresses. There are better ways to handle this but for now we just want to see if the wallet will be a good citizen to servers without UTXO support.