Update Gradle
The SDK was using Gradle 6.8 and has been updated to Gradle 6.9.  Gradle 7.0 appears to be incompatible with the Rust plugin is being used.  Although the application was using Gradle 6.8 previously, the wrapper files hadn't been updated since Gradle 4.6. This commit also corrects that issue.

The sample application is upgraded to Gradle 7.2.