Get tests running to completion
Some tests have been disabled to enable tests to actually complete; there are a number of test failures which need to be addressed.

These tests were failing prior to the refactoring.

The disabled tests can be easily found by searching for the Ignore annotation. They also appear in the Junit results window in Android Studio.