[#373] Add PR and issue templates
This updates our templates to be more consistent with the Secant app repository.

Note that the checklist item for code coverage was removed for now, until we improve the test infrastructure for the SDK.

This also addresses a security reminder, by explicitly stating that one shouldn't run the demo app from a pull request until after reviewing the code changes.