Set signing secrets as command line arguments
It was suggested that GitHub Actions might have an issue with the periods in the signing Gradle properties.  This is an attempt to see if that is the issue.

From Gradle’s perspective `ORG_GRADLE_PROJECT_signing.secretKeyRingFile=` as an environment variable should be equivalent to a command line argument of `-Psigning.secretKeyRingFile=`