Remove `@CriticalNative` usage
This was added to the `TorClient.dispose()` JNI pathway because that
native coded did not require access to `JNIEnv` or `JClass`. However,
it turns out that this annotation was only added in Android 8 for
system usage [0], and was not CTS-tested public API until Android 14.
For whatever reason, in API 27 (Android 8.1) the native method symbol
can't be found, but in API 28 and above it can be.

The performance cost of regular JNI is no longer necessary to worry
about given that we now keep a `TorClient` around long-term as part of
`SdkSynchronizer`, so instead of doing anything complex to preserve it
on API 28+ we just remove the `@CriticalNative` annotation and adjust
the native method's signature to take (and ignore) the extra arguments.

Closes Electric-Coin-Company/zcash-android-wallet-sdk#1587.

[0]: https://developer.android.com/reference/dalvik/annotation/optimization/CriticalNative