Ignore address field in `GetAddressUtxos` response (#1607)
The only thing we were doing with it was passing it through to Rust and
parsing it (panicking if parsing failed). The actual address is fetched
from the UTXO itself, and the Rust code was not doing anything to ensure
the two matched. Given that there seem to be bugs upstream in the address
data, we now just ignore it.