[#1632] Refactor `Account.DEFAULT` (#1645)
* [#1632] Remove `Account.DEFAULT`

* Update Account related APIs

* Refactor balances APIs

* Adopt `getAccounts` in fragment-based Demo app

We have deprecated the old Fragment-based Demo app. See #973. So, the purpose of these changes is purely the buildability of the new Compose-based Demo app.

* Add init to `JniUnifiedSpendingKey.kt`

* Update deprecated Fragment-based Demo app

This part of the Demo app will be removed as part of #973

* Remove deprecated functions from Synchronizer

* Update WalletSnapshot and WalletVM APIs in Demo

+ necessary fixtures changes

* Update newer Compose-based Demo app

* Hide `Synchronizer.createAccount` form public API

As it was, as making it public could bring more requirements on our multi-account support in version 1

* Add all accounts flow API

* Refactor AccountFixture

- We deduplicated the fixture across the related modules and their tests
- Documentation added