rust: Remove mismatched `jboolean` return values from JNI functions
Unlike a C FFI where we need some mechanism to signal an error state,
with JNI thrown exceptions trigger as soon as the JNI function returns.
These functions had no declared return value on the Kotlin side and
thus were not using the returned boolean in any way; this change fixes
the type signatures to match.