Simplify `onScanComplete` trigger resolution
As part of the latest intention of refreshing UTXOs in every `onScanComplete`, we removed the condition for  `refreshAllAccountsUtxos`, `refreshAllBalances`, and `refreshTransactions` as we now want to refresh UTXOs, their balance, and transactions in every block sync loop done. The condition code fragment is old and does not follow the updated SDK state. Moreover, without this optimization, we’re aligned with iOS SDK behavior.