Fix `WalletSummary` being returned as null
A scan progress of `0/0` means there are no shielded notes in the scan
range. This is meant to be treated as 100%. However, on the Rust side of
JNI a top-level filter was dropping the entire `WalletSummary` if the
scan progress denominator was 0 (an artifact of an old approach that no
longer applies). This went unnoticed until the recent Rust crate update,
which included zcash/librustzcash#1773 making 0/0 occur at recovery start.